pyproject.toml
[tool.poetry]
name = "arkanoid"
version = "0.1.0"
description = ""
authors = ["Артур Шарапов <105589311+qaip@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pygame = "^2.4.0"
pyyaml = "^6.0"

[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
app = "arkanoid.main:main"
